scraper/requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
trafilatura==1.12.2
openai==1.57.0
firebase-admin==6.6.0
apscheduler==3.10.4
python-dotenv==1.0.1
feedparser==6.0.11
